pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agency-os"
version = "5.0.0"
authors = [
  { name = "Nelson Arcos", email = "hello@agencyos.dev" }
]
description = "Autonomous AI Agency Operating System — Dev, Marketing, Sales, LeadOps, Analytics, Creative"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "schedule>=1.2",
    "rich>=13.0",
    "httpx>=0.27",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-cov"]

[project.scripts]
agency-os = "kernel.cli:main"
agency = "kernel.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["kernel*", "studios*", "jobs*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
exclude = [
    ".git",
    ".venv",
    "data",
    "logs",
]

[tool.mypy]
exclude = [
    "data/",
    "logs/",
]
